+++ eras/conway/impl/testlib/Test/Cardano/Ledger/Conway/Arbitrary.hs
@@ -434,7 +434,6 @@ instance
       , StakeKeyNotRegisteredDELEG <$> arbitrary
       , StakeKeyHasNonZeroRewardAccountBalanceDELEG <$> arbitrary
       , DRepAlreadyRegisteredForStakeKeyDELEG <$> arbitrary
-      , pure WrongCertificateTypeDELEG
       ]
 
 -- GOVCERT